backend/requirements.txt
fastapi==0.114.*
uvicorn[standard]==0.30.*
python-dotenv==1.0.*
openai==1.54.*
